Modulo de reportes

+++ gestión_de_prestamos_y_finanzas/models/prestamo.py
@@ -141,8 +141,8 @@ class Prestamo(models.Model):
             #amount_per_quota = prestamo.amount_borrowed * tasa_mensual / (1 - (1 + tasa_mensual) ** -prestamo.duration)
             #interest_per_quota = prestamo.amount_borrowed * (prestamo.interest_rate / 100) / 12
             
-            años = float(self.duration/12)
-            cant_cuotas = self.payment_frequency * años
+            años = int(self.duration/12)
+            cant_cuotas = int(self.payment_frequency) * años
             At = self.amount_borrowed / cant_cuotas
             n = 0
             for quta in range(1, cant_cuotas + 1):
